container and file refactor

+++ recsys/core/dal/repo.py
@@ -11,7 +11,7 @@
 # URL        : https://github.com/john-james-ai/Recommender-Systems                                #
 # ------------------------------------------------------------------------------------------------ #
 # Created    : Monday November 14th 2022 01:27:04 am                                               #
-# Modified   : Thursday November 24th 2022 04:42:50 am                                             #
+# Modified   : Friday November 25th 2022 05:33:50 pm                                               #
 # ------------------------------------------------------------------------------------------------ #
 # License    : MIT License                                                                         #
 # Copyright  : (c) 2022 John James                                                                 #
@@ -19,21 +19,13 @@
 """Data Repository Module"""
 import logging
 import os
-from dotenv import load_dotenv
 from abc import ABC, abstractmethod
 from typing import Any
-from functools import wraps
 import shutil
 
-
-from dependency_injector import containers, providers
-from dependency_injector.wiring import inject
-
 from recsys.core.dal.dataset import Dataset
-from recsys.core.dal.database import Database
 from recsys.core.dal.registry import DatasetRegistry
 from recsys.core.services.io import IOService
-from recsys.config.base import REPO_DIRS, FilesetInput, DatasetInput
 
 # ------------------------------------------------------------------------------------------------ #
 logger = logging.getLogger(__name__)
@@ -64,27 +56,29 @@ class Repo(ABC):
 class DatasetRepo(Repo):
     """Dataset Repository"""
 
-    __registry = {}
-
-    def __init__(self, io: IOService, registry: DatasetRegistry, file_format: str = "pkl") -> None:
+    def __init__(
+        self, directory: str, io: IOService, registry: DatasetRegistry, file_format: str
+    ) -> None:
+        self._directory = directory
         self._io = io
-        DatasetRepo.__registry = registry
+        self._registry = registry
         self._file_format = file_format
 
-        self._directory = self._get_directory()
-
-        self._validate()
-
     def __len__(self) -> int:
-        return len(DatasetRepo.__registry)
+        return len(self._registry)
 
     @property
     def directory(self) -> str:
         return self._directory
 
     def reset(self) -> None:
-        DatasetRepo.__registry.reset()
-        shutil.rmtree(self._directory, ignore_errors=True)
+        msg = "Are you SURE. This will delete EVERYTHING in this repository. If you are sure, type 'YES, I AM SURE': "
+        sure = input(msg)
+        if sure == "YES, I AM SURE":
+            logger.info(f"\n\nRESETTING REPOSITORY at {self._directory}.")
+            self._registry.reset()
+            shutil.rmtree(self._directory, ignore_errors=True)
+            logger.info(f"\n\nREPOSITORY at {self._directory}, RESET.")
 
     def add(self, dataset: Dataset) -> None:
         """Adds a Dataset to the repo.
@@ -92,11 +86,11 @@ class DatasetRepo(Repo):
         Args:
             dataset (Dataset): The Dataset object
         """
-        while DatasetRepo.__registry.version_exists(dataset):
+        while self._registry.version_exists(dataset):
             dataset.version = dataset.version + 1
 
         dataset.filepath = self._get_filepath(dataset)
-        dataset = DatasetRepo.__registry.add(dataset)
+        dataset = self._registry.add(dataset)
         os.makedirs(os.path.dirname(dataset.filepath), exist_ok=True)
         self._io.write(filepath=dataset.filepath, data=dataset)
         return dataset
@@ -108,22 +102,21 @@ class DatasetRepo(Repo):
             id (int): ID of the dataset object
         """
         try:
-            registration = DatasetRepo.__registry.get(id)
+            registration = self._registry.get(id)
             return self._io.read(registration["filepath"])
         except FileNotFoundError:
             msg = f"No Dataset with id = {id} exists."
             logger.error(msg)
             raise FileNotFoundError(msg)
 
-    def find_dataset(self, name: str, env: str = None, stage: str = None):
+    def find_dataset(self, name: str, stage: str = None):
         """Returns a Dataframe of Dataset objects that match the criteria
 
         Args:
             name (str): Required name of Dataset.
-            env (str): Optional, unless stage is provided. One of 'test', 'dev', or 'prod'.
             stage (str): Optional, one of 'raw', 'interim', or 'cooked'.
         """
-        return DatasetRepo.__registry.find_dataset(name=name, env=env, stage=stage)
+        return self._registry.find_dataset(name=name, stage=stage)
 
     def remove(self, id: int, ignore_errors=False) -> None:
         """Removes a Dataset object from the registry and file system.
@@ -135,9 +128,9 @@ class DatasetRepo(Repo):
         Raises: FileNotFoundError (if ignore_errors = False)
         """
         try:
-            registration = DatasetRepo.__registry.get(id)
+            registration = self._registry.get(id)
             os.remove(registration["filepath"])
-            DatasetRepo.__registry.remove(id)
+            self._registry.remove(id)
 
         except FileNotFoundError:
             if ignore_errors:
@@ -152,15 +145,15 @@ class DatasetRepo(Repo):
         Args:
             dataset (Dataset): The dataset to check.
         """
-        return DatasetRepo.__registry.exists(id)
+        return self._registry.exists(id)
 
     def version_exists(self, dataset: Dataset) -> bool:
         """Checks whether the dataset exists.
         Args:
             dataset (Dataset): The dataset to check.
         """
-        return DatasetRepo.__registry.version_exists(
-            name=dataset.name, env=dataset.env, stage=dataset.stage, version=dataset.version
+        return self._registry.version_exists(
+            name=dataset.name, stage=dataset.stage, version=dataset.version
         )
 
     def print(self, id: int) -> None:
@@ -171,7 +164,7 @@ class DatasetRepo(Repo):
 
         """
         try:
-            registration = DatasetRepo.__registry.get(id=id)
+            registration = self._registry.get(id=id)
             dataset = Dataset(**registration)
             print(dataset)
         except FileNotFoundError:
@@ -181,35 +174,12 @@ class DatasetRepo(Repo):
 
     def print_registry(self) -> None:
         """Prints list of dataset names in repository."""
-        print(DatasetRepo.__registry.get_all())
-
-    def _validate(self) -> None:
-        if not isinstance(self._io, (type(IOService), IOService)):
-            msg = f"{type(self._io)} is not a valid io service."
-            logger.error(msg)
-            raise ValueError(msg)
-        if not isinstance(DatasetRepo.__registry, DatasetRegistry):
-            msg = f"{type(DatasetRepo.__registry)} is not a valid registry object."
-            logger.error(msg)
-            raise ValueError(msg)
-
-    def _get_directory(self) -> str:
-        """Returns the repo directory based upon the current environment."""
-        load_dotenv()
-        ENV = os.getenv("ENV")
-        try:
-            return REPO_DIRS["data"].get(ENV)
-        except KeyError:
-            msg = "The current environment, specified by the 'ENV' variable in the .env file, is not supported."
-            logger.error(msg)
-            raise ValueError(msg)
+        print(self._registry.get_all())
 
     def _get_filepath(self, dataset: Dataset) -> str:
-        """Formats a filename using the Dataset name, environment, stage, and version."""
+        """Formats a filename using the Dataset name, stage, and version."""
         filename = (
             dataset.name
-            + "_"
-            + dataset.env
             + "_"
             + dataset.stage
             + "_v"
@@ -217,78 +187,4 @@ class DatasetRepo(Repo):
             + "."
             + self._file_format
         )
-        return os.path.join(self._directory, dataset.env, dataset.stage, filename)
-
-
-# ------------------------------------------------------------------------------------------------ #
-#                                       CONTAINER                                                  #
-# ------------------------------------------------------------------------------------------------ #
-
-
-class Container(containers.DeclarativeContainer):
-
-    # Services
-    db = providers.Singleton(Database, database="data")
-
-    registry = providers.Singleton(DatasetRegistry, database=db)
-
-    repo = providers.Singleton(DatasetRepo, io=IOService, registry=registry, file_format="pkl")
-
-
-# ------------------------------------------------------------------------------------------------ #
-#                                   DECORATOR FUNCTIONS                                            #
-# ------------------------------------------------------------------------------------------------ #
-def repository(func):
-    @wraps(func)
-    @inject
-    def wrapper(self, *args, **kwargs):
-
-        container = Container()
-        repo = container.repo()
-
-        datasets = {}
-        dataset = None
-
-        # Handle input
-        if hasattr(self, "input_params"):
-            if isinstance(self.input_params, FilesetInput):
-                pass  # This will be handled by the operator.
-            elif isinstance(self.input_params, DatasetInput):
-                dataset = repo.get(self.input_params.id)
-                setattr(self, "input_dataset", dataset.data)
-            elif isinstance(self.input_params, dict):
-                for k, v in self.input_params.items():
-                    if isinstance(v, FilesetInput):
-                        pass  # Again, handled by operator
-                    elif isinstance(v, DatasetInput):
-                        datasets[v.name] = repo.get(v.id)
-                    else:
-                        msg = f"{self.input_params} is unrecognized input."
-                        logger.error(msg)
-                        raise TypeError(msg)
-                if len(datasets) > 0:
-                    setattr(func, "input_dataset", datasets)
-
-        # Execute wrapped method.
-        result = func(self, *args, **kwargs)
-
-        # Handle Results
-
-        results = {}
-
-        def store_result(result) -> None:
-            if isinstance(result, dict):
-                for k, v in result.items():
-                    results[k] = repo.add(v)
-                return results
-            elif isinstance(result, Dataset):
-                return repo.add(result)
-            else:
-                msg = "Result was not a Dataset or dictionary object."
-                logger.error(msg)
-                raise TypeError(msg)
-
-        return store_result(result)
-        repo.print_registry()
-
-    return wrapper
+        return os.path.join(self._directory, dataset.stage, filename)
